+++ src/commands/go/goFish.js
@@ -1,19 +1,7 @@
+import { allFishEmojis, allFishNames, allDBNames } from '../../constants';
+
 const goFish = (stopClient, msg, cmd, subcmd, richEmbed, fishList) => {
   if (subcmd === 'fish') {
-
-    // Arrays of fish names, emojis names, and db names
-    const fishEmojis = [':fish:', ':fish_cake:', ':fishing_pole_and_fish:', ':tropical_fish:', ':blowfish:'];
-    const fishNames = ['Fish', 'Fish Cake', 'Fish on a Fishing Pole', 'Tropical Fish', 'Blowfish'];
-    const dbNames = ['fish', 'cake', 'fishpole', 'tropical', 'blowfish'];
-
-    const rareFishEmojis = [':whale:', ':whale2:', ':dolphin:', ':octopus:', ':unicorn:'];
-    const dbRare = ['cutewhale', 'bluewhale', 'dolphin', 'octopus', 'unicorn'];
-    const rareFishNames = ['Cute Whale', 'Blue Whale', 'Dolphin', 'Octopus', 'Unicorn'];
-
-    const allFishEmojis = fishEmojis.concat(fishEmojis).concat(fishEmojis).concat(rareFishEmojis);
-    const allFishNames = fishNames.concat(fishNames).concat(fishNames).concat(rareFishNames);
-    const allDBNames = dbNames.concat(dbNames).concat(dbNames).concat(dbRare);
-
     // Reset to empty array on every "!go fish" message
     const fishCaught = [];
 
@@ -44,7 +32,7 @@ const goFish = (stopClient, msg, cmd, subcmd, richEmbed, fishList) => {
 
   **${fishList['bluewhale']}**  x  :whale2:      **${fishList['dolphin']}**  x  :dolphin:      **${fishList['octopus']}**  x  :octopus:
   `);
-    msg.channel.send({embed});
+    msg.channel.send({ embed });
   }
 }